wgpu_sync::atomic

Type Alias AtomicPtr
·

pub type AtomicPtr<T> = Atomic<*mut T>;

A raw pointer type which can be safely shared between threads.

This type has the same size and bit validity as a *mut T.

Note: This type is only available on platforms that support atomic loads and stores of pointers. Its size depends on the target pointer’s size.

Aliased Type

#[repr(C)]
pub struct AtomicPtr<T> { /* private fields */ }